refactor[frontend]: edit importing

+++ views/sniffing/uart_config.py
@@ -16,8 +16,8 @@ class UartConfigurations(QDialog):
         self.setGeometry(100, 100, 400, 300)
 
         screen_geometry = QApplication.desktop().availableGeometry()
-        x = (screen_geometry.width() - self.width()) / 2
-        y = (screen_geometry.height() - self.height()) / 2
+        x = int((screen_geometry.width() - self.width()) / 2)
+        y = int((screen_geometry.height() - self.height()) / 2)
         self.move(x, y)
 
         self.uart_settings()
